requirements.txt
pandas
plotly
sklearn.model_selection
matplotlib
numpy
typing